Added tournament state.

+++ pokertournament-api/src/main/java/com/pilon/pokertournament/tournaments/TournamentBreak.java
@@ -31,7 +31,7 @@ public class TournamentBreak {
 
     @Column
     @NotNull
-    private Short afterLevel;
+    private Integer afterLevel;
 
     @Column
     @NotNull
